Add progress bar to pages
