+++ package.json
@@ -16,6 +16,7 @@
     "react-bootstrap": "^2.0.0-beta.4",
     "react-circular-progressbar": "^2.0.4",
     "react-dom": "^17.0.2",
+    "react-moment": "^1.1.1",
     "react-redux": "^7.2.4",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
