+++ src/Pages/Show.jsx
@@ -3,16 +3,25 @@ import React, { useEffect, useState } from "react";
 import { Card, Col, Container } from "react-bootstrap";
 import { Link, useHistory, useParams } from "react-router-dom";
 import CircularProgress from "@material-ui/core/CircularProgress";
+import Moment from "react-moment";
+import dateHandle from "../PureFunctions/time";
+import final from "../PureFunctions/date";
+import { useSelector } from "react-redux";
+import Progress from "../components/Progress";
 const Show = () => {
   const [data, setData] = useState([]);
   const { id } = useParams();
   const { name } = useParams();
   const history = useHistory();
 
+  const [progress, setProgress] = React.useState(0);
+  const measurments = useSelector((state) => state.measurments);
+
   const fetchData = async () => {
     try {
       const data = await axios.get(`http://localhost:3001/measurments/${id}`);
       setData(data.data);
+      setProgress(data.data.number);
     } catch (e) {
       console.log(e);
     }
@@ -34,6 +43,7 @@ const Show = () => {
   };
   return (
     <Container className="cont">
+      <Progress number={progress} time="Today" />
       <Col>
         <h1 className="text-center">{data.id && name}</h1>
         {data.id && (
@@ -45,16 +55,13 @@ const Show = () => {
             >
               <Card>
                 <Card.Body className="d-flex justify-content-between">
-                  <Card.Text className="my-auto text-decoration-none link-dark">
-                    {name}
-                  </Card.Text>
-                  <Card.Text className="my-auto text-decoration-none link-dark">
-                    {data.number}
-                  </Card.Text>
-                  <Card.Text className="my-auto text-decoration-none link-dark">
-                    {data.date}
-                  </Card.Text>
                   <CircularProgress variant="determinate" value={data.number} />
+                  <Moment className="text-dark" format="LL">
+                    {data.created_at}
+                  </Moment>
+                  <Card.Text className="my-auto text-decoration-none link-dark">
+                    {data.number} {name.charAt(0).toUpperCase() + name.slice(1)}
+                  </Card.Text>
                 </Card.Body>
               </Card>
             </Link>
